% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export_widget.R
\name{export_widget}
\alias{export_widget}
\title{Export a htmlwidget SVG as PNG}
\usage{
export_widget(selector = NULL, width = 0, height = 0)
}
\description{
Often we want to export a SVG htmlwidget as a static PNG.  This little helper
widget will add a button on your widget to do just this.
}

